requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
Flask==3.1.1
fonttools==4.58.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jira==3.8.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
slack_sdk==3.35.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
